requirements.txt
sqlalchemy
pymssql
pymysql
pypattyrn
cryptography
fastapi
Pillow
requests
